test::with_timer
