+++ lumen_runtime/src/otp/erlang/read_timer_1/test.rs
@@ -6,8 +6,7 @@ use liblumen_alloc::erts::term::prelude::*;
 use crate::otp::erlang;
 use crate::otp::erlang::read_timer_1::native;
 use crate::process;
-use crate::test::{has_message, receive_message, timeout_message};
-use crate::time::Milliseconds;
+use crate::test::{has_message, timeout_message};
 use crate::timer;
 
 #[test]
