+++ lumen_runtime/src/test.rs
@@ -13,14 +13,20 @@ pub mod strategy;
 pub use self::proptest::*;
 
 use std::convert::TryInto;
+use std::sync::{Arc, Barrier};
+use std::thread;
+use std::time::Duration;
 
 use liblumen_alloc::atom;
 use liblumen_alloc::erts::exception;
 use liblumen_alloc::erts::process::{Process, Status};
 use liblumen_alloc::erts::term::prelude::*;
 
-use crate::process::SchedulerDependentAlloc;
+use crate::otp::erlang;
+use crate::process::{self, SchedulerDependentAlloc};
 use crate::scheduler::with_process;
+use crate::time::Milliseconds;
+use crate::timer;
 
 pub fn assert_exits<F: Fn(Option<Term>)>(
     process: &Process,
@@ -80,6 +86,87 @@ pub fn badarity_reason(process: &Process, function: Term, args: Term) -> Term {
     let fun_args = process.tuple_from_slice(&[function, args]).unwrap();
 
     process.tuple_from_slice(&[tag, fun_args]).unwrap()
+}
+
+pub fn timeout_after_half(milliseconds: Milliseconds, barrier: &Barrier) {
+    thread::sleep(Duration::from_millis(milliseconds / 2 + 1));
+    timer::timeout();
+    barrier.wait();
+}
+
+pub fn wait_for_completion(barrier: &Barrier) {
+    barrier.wait();
+}
+
+pub fn wait_for_message(barrier: &Barrier) {
+    barrier.wait();
+}
+
+pub fn with_timer<F>(f: F)
+where
+    F: FnOnce(u64, &Barrier, Term, &Process) -> (),
+{
+    let same_thread_process_arc = process::test(&process::test_init());
+    let milliseconds: u64 = 100;
+
+    // no wait to receive implemented yet, so use barrier for signalling
+    let same_thread_barrier = Arc::new(Barrier::new(2));
+
+    let different_thread_same_thread_process_arc = Arc::clone(&same_thread_process_arc);
+    let different_thread_barrier = same_thread_barrier.clone();
+
+    let different_thread = thread::spawn(move || {
+        let different_thread_process_arc = process::test(&different_thread_same_thread_process_arc);
+        let same_thread_pid = different_thread_same_thread_process_arc.pid();
+
+        let timer_reference = erlang::start_timer_3::native(
+            different_thread_process_arc.clone(),
+            different_thread_process_arc.integer(milliseconds).unwrap(),
+            same_thread_pid.into(),
+            Atom::str_to_term("different"),
+        )
+        .unwrap();
+
+        erlang::send_2::native(
+            &different_thread_process_arc,
+            same_thread_pid.into(),
+            different_thread_process_arc
+                .tuple_from_slice(&[Atom::str_to_term("timer_reference"), timer_reference])
+                .unwrap(),
+        )
+        .expect("Different thread could not send to same thread");
+
+        wait_for_message(&different_thread_barrier);
+        timeout_after_half(milliseconds, &different_thread_barrier);
+        timeout_after_half(milliseconds, &different_thread_barrier);
+
+        // stops Drop of scheduler ID
+        wait_for_completion(&different_thread_barrier);
+    });
+
+    wait_for_message(&same_thread_barrier);
+
+    let timer_reference_tuple =
+        receive_message(&same_thread_process_arc).expect("Cross-thread receive failed");
+
+    let timer_reference = erlang::element_2::native(
+        same_thread_process_arc.integer(2).unwrap(),
+        timer_reference_tuple,
+    )
+    .unwrap();
+
+    f(
+        milliseconds,
+        &same_thread_barrier,
+        timer_reference,
+        &same_thread_process_arc,
+    );
+
+    wait_for_completion(&same_thread_barrier);
+
+    different_thread
+        .join()
+        .expect("Could not join different thread");
 }
 
 pub fn without_timer_returns_false(native: fn(&Process, Term) -> exception::Result<Term>) {
